requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
cloudinary==1.29.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==4.0.4
django-cloudinary-storage==0.3.0
django-environ==0.10.0
django-import-export==3.2.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.1.1
psycopg2-binary==2.9.3
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tablib==3.4.0
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.1.0
xlrd==2.0.1
xlwt==1.3.0
